fix(transfer-event-builder): split BuilderServices + TransferDirection out
The rebase over #774 (counterpartyAddress) put TransferEventBuilder.swift
at 419 lines vs SwiftLint's 400-line file_length cap. Extract two helper
types into sibling files: `BuilderServices` (the per-cycle DI bundle from
this PR) and `TransferDirection` (the lowercased-from/to direction enum
that pre-dates this PR but lives next to where it's used).

No behavior change. No baseline edits.
